refactor(support): comply with clean architecture and api standards
Introduced HealthService to decouple UseCase from Repository. standardized API response using ResponseHelper. moved HealthStatus to @sgcv2/shared. and cleaned up dead code in helpers.
