+++ backend/src/modules/support/application/get-health.use-case.ts
@@ -1,29 +1,16 @@
 import { inject, injectable } from 'inversify';
-import { HealthCheckRepository } from '../domain/repositories/health-check.repository';
+import { HealthService } from '../domain/services/health.service';
 import { TYPES } from '@modules/support/di/types';
-
-export interface HealthStatus {
-  status: 'ok' | 'error';
-  timestamp: string;
-  environment: string;
-  database: 'connected' | 'disconnected';
-}
+import { HealthStatus } from '@sgcv2/shared';
 
 @injectable()
 export class GetHealthUseCase {
   constructor(
-    @inject(TYPES.HealthCheckRepository)
-    private readonly repository: HealthCheckRepository
-  ) {}
+    @inject(TYPES.HealthService)
+    private readonly healthService: HealthService
+  ) { }
 
   async execute(): Promise<HealthStatus> {
-    const isDbConnected = await this.repository.checkDatabase();
-
-    return {
-      status: isDbConnected ? 'ok' : 'error',
-      timestamp: new Date().toISOString(),
-      environment: process.env.NODE_ENV || 'development',
-      database: isDbConnected ? 'connected' : 'disconnected',
-    };
+    return this.healthService.checkHealth();
   }
 }
