+++ backend/src/modules/support/di/types.ts
@@ -1,5 +1,6 @@
 export const TYPES = {
   HealthCheckRepository: Symbol.for('HealthCheckRepository'),
+  HealthService: Symbol.for('HealthService'),
   GetHealthUseCase: Symbol.for('GetHealthUseCase'),
   HealthController: Symbol.for('HealthController'),
   SupportRoutes: Symbol.for('SupportRoutes'),
