+++ backend/src/modules/support/domain/services/health.service.ts
@@ -1,0 +1,5 @@
+import { HealthStatus } from '@sgcv2/shared';
+
+export interface HealthService {
+  checkHealth(): Promise<HealthStatus>;
+}
