+++ backend/tests/integration/health.test.ts
@@ -21,8 +21,9 @@ describe('Health Check Integration Test', () => {
     const response = await request(app).get('/api/health');
 
     expect(response.status).toBe(200);
-    expect(response.body).toHaveProperty('status', 'ok');
-    expect(response.body).toHaveProperty('database', 'connected');
-    expect(response.body).toHaveProperty('timestamp');
+    expect(response.body.success).toBe(true);
+    expect(response.body.data).toHaveProperty('status', 'ok');
+    expect(response.body.data).toHaveProperty('database', 'connected');
+    expect(response.body.data).toHaveProperty('timestamp');
   });
 });
